requirements.txt
jupyter
numpy
catboost
seaborn
plotly
# serve
fastapi
uvicorn==0.17.6
# experiments
mlflow-skinny==2.16.2
hyperopt
boto3==1.35
# retrieval
python-dotenv==1.0.1
scikit-learn
umap-learn
PyYAML==6.0.1
# data processing
polars==1.15.0
pyarrow==18.1.0
# bot
python-telegram-bot==21.8
openai==1.57.0

backoff==2.2.1
pyarrow==18.1.0
torch==2.3.1
transformers==4.41.2
sentence-transformers==2.7.0
torchvision
nltk==3.6.7
gensim
implicit
google-genai
label-studio
catalyst
chromadb
langchain
langchain-text-splitters
langchain_huggingface